CompleteProject/requirements_mysql.txt
dash==2.14.0
dash-core-components==2.0.0
dash-html-components==2.0.0
pandas==2.0.0
plotly==5.18.0
numpy==1.24.0
mysql-connector-python==8.0.33
gunicorn==21.2.0
